Move contents of forced_steps.py into playground.py and delete forced_steps.py
